WIP: Continued clean up. Delete everything but broadcast for now.
